requirements.txt
spotipy==2.24.0
requests==2.31.0
numpy==1.26.2
pyyaml==6.0.1
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78